# more optimizations