@aminnj 2021-09-12
The majority of `interped_data` is now spent in `real_data = ntoh.(reinterpret(T, rawdata))`.