@tamasgal 2021-09-12
I am wondering if we can do it easily in-place...