@aminnj 2021-09-12
I played around a bit. It got messy. I wrapped the reinterpreted array (of type `Base.ReinterpretArray{...}`) to get a Vector{T} and then do `ntoh` in place.

before             |  after
:-------------------------:|:-------------------------:
![image](https://user-images.githubusercontent.com/5760027/132978515-a273cef3-de8f-427a-a025-cee3c1410b3c.png)  |  ![image](https://user-images.githubusercontent.com/5760027/132978521-2d72e0f6-ff28-4568-bfa4-039bd56ac0d0.png)

And it works in the REPL with the loop in the original post and gives another 4-5% speedup. But unit tests fail Probably I shouldn't be using `unsafe_*` without knowing the consequences. I even tried some GC protection...
```julia
# real_data = ntoh.(reinterpret(T, rawdata))
n = length(rawdata)÷_size
ptr = pointer(reinterpret(T, rawdata))
real_data = unsafe_wrap(Array, ptr, (n,))
GC.@preserve rawdata real_data begin
    real_data .= ntoh.(real_data)
    vov = VectorOfVectors(real_data, offset, ArraysOfArrays.no_consistency_checks)
end
return vov
```
but no success.

If I run the unit tests by hand in the REPL, they work. If I put printouts right before certain failures, they suddenly work. Seems like GC?