@Moelf 2021-09-12
see https://github.com/aminnj/UnROOT.jl/pull/2