@aminnj 2021-09-12
I tried that PR but it was the same speed as before (`func1` -- master), so there's clearly something different about the `unsafe_wrap` I did above (`func2`) and yours (`func3`).
```julia
julia> function func1(rawdata)
           T = Int32
           real_data = ntoh.(reinterpret(T, rawdata))
           return real_data
       end

julia> function func2(rawdata)
          T = Int32
          _size = sizeof(T)
          n = length(rawdata)÷_size
          ptr = pointer(reinterpret(T, rawdata))
          real_data = unsafe_wrap(Array, ptr, (n,))
          GC.@preserve rawdata real_data begin
              real_data .= ntoh.(real_data)
           end
           real_data
       end

julia> function func3(rawdata)
           T = Int32
           _eltype = T
           dp = length(rawdata)
           _size = sizeof(T)
           real_data = GC.@preserve rawdata ntoh.(unsafe_wrap(Array, Ptr{_eltype}(pointer(rawdata)), dp÷_size))
           return real_data
       end

julia> @benchmark func1($x) setup=begin x = rand(UInt8,1_000_000) end
 Range (min … max):  164.473 μs …   6.542 ms  ┊ GC (min … max):  0.00% … 89.25%
 Time  (median):     592.193 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   602.361 μs ± 682.423 μs  ┊ GC (mean ± σ):  14.76% ± 12.02%

julia> @benchmark func2($x) setup=begin x = rand(UInt8,1_000_000) end
 Range (min … max):  28.773 μs … 222.436 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     35.880 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   42.340 μs ±  15.488 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

julia> @benchmark func3($x) setup=begin x = rand(UInt8,1_000_000) end
 Range (min … max):   81.082 μs …   6.492 ms  ┊ GC (min … max):  0.00% … 86.11%
 Time  (median):     470.794 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   475.018 μs ± 613.650 μs  ┊ GC (mean ± σ):  14.62% ± 10.77%

```
If we can get `func2` to work safely (so that it works in the REPL and unit tests and everywhere else) that would be super nice!