@aminnj 2021-09-14
Should we merge the changes in the original post into `master`? Then we can worry about additional `unsafe_*` optimizations separately, to avoid conflating the two speedups into one.